+++ sys/net/gnrc/pktbuf/Makefile
@@ -1,0 +1,3 @@
+MODULE = gnrc_pktbuf
+
+include $(RIOTBASE)/Makefile.base
